fix clang compile errors printing size_t values

+++ media/base/shell_buffer_factory.cc
@@ -143,7 +143,7 @@ bool ShellBufferFactory::AllocateBuffer(
     } else {
       // Alright, we have to wait, enqueue the buffer and size.
       DLOG(INFO) << base::StringPrintf(
-          "deferred allocation of %d bytes", size);
+          "deferred allocation of %d bytes", (int)size);
       filter_graph_log->LogEvent(kObjectIdBufferFactory,
                                  kEventBufferAllocationDeferred);
       pending_allocs_.push_back(
@@ -179,7 +179,8 @@ uint8* ShellBufferFactory::AllocateNow(size_t size) {
   }
 
   if (!bytes) {
-    DLOG(ERROR) << base::StringPrintf("Failed to allocate %d bytes!", size);
+    DLOG(ERROR) << base::StringPrintf("Failed to allocate %d bytes!",
+                                      (int)size);
   }
 
   return bytes;
